requirements.rag.txt
# RAG Dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.0.0
numpy>=1.24.0
